+++ example_scripts/print.js
@@ -1,0 +1,3 @@
+LootJS.lootTables(event => {
+    event.getLootTable("minecraft:chests/simple_dungeon").print()
+})