+++ src/main/java/com/almostreliable/lootjs/core/entry/EmptyLootEntry.java
@@ -1,5 +1,6 @@
 package com.almostreliable.lootjs.core.entry;
 
+import com.almostreliable.lootjs.util.DebugInfo;
 import net.minecraft.world.level.storage.loot.entries.EmptyLootItem;
 import net.minecraft.world.level.storage.loot.entries.LootPoolSingletonContainer;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
@@ -21,5 +22,11 @@ public class EmptyLootEntry extends AbstractSimpleLootEntry<EmptyLootItem> imple
     public EmptyLootEntry addCondition(LootItemCondition condition) {
         getConditions().add(condition);
         return this;
+    }
+
+    @Override
+    public void collectDebugInfo(DebugInfo info) {
+        info.add("% Empty");
+        super.collectDebugInfo(info);
     }
 }