+++ src/main/java/com/almostreliable/lootjs/loot/LootEntryList.java
@@ -66,9 +66,14 @@ public class LootEntryList extends ListHolder<LootEntry, LootPoolEntryContainer>
 
         info.add("% Entries: [");
         info.push();
+        info.add("{");
+        info.push();
         for (LootEntry entry : this) {
             entry.collectDebugInfo(info);
         }
+
+        info.pop();
+        info.add("}");
         info.pop();
         info.add("]");
     }